chore(docs): Optimized array indexing in the log monitor widget's rendering logic
# Project State

## Current Focus
Optimized array indexing in the log monitor widget's rendering logic

## Context
The change improves performance by simplifying the array indexing calculation in the log monitor widget's rendering logic.

## Completed
- [x] Removed redundant cast in array index calculation
- [x] Simplified the target index calculation for better readability and potential performance

## In Progress
- [x] No active work in progress beyond this change

## Blockers
- None identified

## Next Steps
1. Verify no regression in log monitor rendering
2. Consider additional performance optimizations in related widgets
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -243,7 +243,7 @@ impl Widget for LogMonitor {
             let src_x = la.x + col as u16;
             let src_y = la.y + row as u16;
             if src_y < area.height - 1 && src_x < area.width - 1 {
-                let target = (src_y as usize * w + src_x as usize) as usize;
+                let target = src_y as usize * w + src_x as usize;
                 if target < p.cells.len() {
                     p.cells[target] = c.clone();
                 }
